+++ tap_shopify/__init__.py
@@ -4,6 +4,7 @@ import datetime
 import json
 import time
 import math
+import copy
 
 import pyactiveresource
 import shopify
@@ -12,17 +13,27 @@ from singer import utils
 from singer import metadata
 from singer import Transformer
 from tap_shopify.context import Context
+from tap_shopify.exceptions import ShopifyError
+from tap_shopify.streams.base import shopify_error_handling, get_request_timeout
 import tap_shopify.streams # Load stream objects into Context
 
 REQUIRED_CONFIG_KEYS = ["shop", "api_key"]
 LOGGER = singer.get_logger()
+SDC_KEYS = {'id': 'integer', 'name': 'string', 'myshopify_domain': 'string'}
 
+@shopify_error_handling
 def initialize_shopify_client():
     api_key = Context.config['api_key']
     shop = Context.config['shop']
     version = '2021-07'
     session = shopify.Session(shop, version, api_key)
     shopify.ShopifyResource.activate_session(session)
+
+    # set request timeout
+    shopify.Shop.set_timeout(get_request_timeout())
+
+    # Shop.current() makes a call for shop details with provided shop and api_key
+    return shopify.Shop.current().attributes
 
 def get_abs_path(path):
     return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
@@ -68,7 +79,14 @@ def load_schema_references():
 
     return refs
 
+def add_synthetic_key_to_schema(schema):
+    for k in SDC_KEYS:
+        schema['properties']['_sdc_shop_' + k] = {'type': ["null", SDC_KEYS[k]]}
+    return schema
+
 def discover():
+    initialize_shopify_client() # Checking token in discover mode
+
     raw_schemas = load_schemas()
     streams = []
 
@@ -79,12 +97,20 @@ def discover():
 
         stream = Context.stream_objects[schema_name]()
 
+        # resolve_schema_references() is changing value of passed refs.
+        # Customer is a stream and it's a nested field of orders and abandoned_checkouts streams
+        # and those 3 _sdc fields are also added inside nested field customer for above 2 stream
+        # so create a copy of refs before passing it to resolve_schema_references().
+        refs_copy = copy.deepcopy(refs)
+        catalog_schema = add_synthetic_key_to_schema(
+            singer.resolve_schema_references(schema, refs_copy))
+
         # create and add catalog entry
         catalog_entry = {
             'stream': schema_name,
             'tap_stream_id': schema_name,
-            'schema': singer.resolve_schema_references(schema, refs),
-            'metadata' : get_discovery_metadata(stream, schema),
+            'schema': catalog_schema,
+            'metadata': get_discovery_metadata(stream, schema),
             'key_properties': stream.key_properties,
             'replication_key': stream.replication_key,
             'replication_method': stream.replication_method
@@ -106,8 +132,10 @@ def shuffle_streams(stream_name):
     bottom_half = Context.catalog["streams"][:matching_index]
     Context.catalog["streams"] = top_half + bottom_half
 
+# pylint: disable=too-many-locals
 def sync():
-    initialize_shopify_client()
+    shop_attributes = initialize_shopify_client()
+    sdc_fields = {"_sdc_shop_" + x: shop_attributes[x] for x in SDC_KEYS}
 
     # Emit all schemas first so we have them for child streams
     for stream in Context.catalog["streams"]:
@@ -144,7 +172,9 @@ def sync():
                 extraction_time = singer.utils.now()
                 record_schema = catalog_entry['schema']
                 record_metadata = metadata.to_map(catalog_entry['metadata'])
-                rec = transformer.transform(rec, record_schema, record_metadata)
+                rec = transformer.transform({**rec, **sdc_fields},
+                                            record_schema,
+                                            record_metadata)
                 singer.write_record(stream_id,
                                     rec,
                                     time_extracted=extraction_time)
@@ -160,25 +190,41 @@ def sync():
 
 @utils.handle_top_exception(LOGGER)
 def main():
-
-    # Parse command line arguments
-    args = utils.parse_args(REQUIRED_CONFIG_KEYS)
-
-    # If discover flag was passed, run discovery mode and dump output to stdout
-    if args.discover:
-        catalog = discover()
-        print(json.dumps(catalog, indent=2))
-    # Otherwise run in sync mode
-    else:
-        Context.tap_start = utils.now()
-        if args.catalog:
-            Context.catalog = args.catalog.to_dict()
-        else:
-            Context.catalog = discover()
+    try:
+        # Parse command line arguments
+        args = utils.parse_args(REQUIRED_CONFIG_KEYS)
 
         Context.config = args.config
         Context.state = args.state
-        sync()
+
+        # If discover flag was passed, run discovery mode and dump output to stdout
+        if args.discover:
+            catalog = discover()
+            print(json.dumps(catalog, indent=2))
+        # Otherwise run in sync mode
+        else:
+            Context.tap_start = utils.now()
+            if args.catalog:
+                Context.catalog = args.catalog.to_dict()
+            else:
+                Context.catalog = discover()
+
+            sync()
+    except pyactiveresource.connection.ResourceNotFound as exc:
+        raise ShopifyError(exc, 'Ensure shop is entered correctly') from exc
+    except pyactiveresource.connection.UnauthorizedAccess as exc:
+        raise ShopifyError(exc, 'Invalid access token - Re-authorize the connection') \
+            from exc
+    except pyactiveresource.connection.ConnectionError as exc:
+        msg = ''
+        try:
+            body_json = exc.response.body.decode()
+            body = json.loads(body_json)
+            msg = body.get('errors')
+        finally:
+            raise ShopifyError(exc, msg) from exc
+    except Exception as exc:
+        raise ShopifyError(exc) from exc
 
 if __name__ == "__main__":
     main()