+++ tests/base.py
@@ -6,7 +6,9 @@ import unittest
 import os
 from datetime import datetime as dt
 from datetime import timezone as tz
-
+import dateutil.parser
+import pytz
+from datetime import timedelta
 from tap_tester import connections, menagerie, runner
 
 
@@ -24,6 +26,8 @@ class BaseTapTest(unittest.TestCase):
     INCREMENTAL = "INCREMENTAL"
     FULL = "FULL_TABLE"
     START_DATE_FORMAT = "%Y-%m-%dT%H:%M:%SZ"
+    BOOKMARK_COMPARISON_FORMAT = "%Y-%m-%dT00:00:00+00:00"
+    DEFAULT_RESULTS_PER_PAGE = 175
 
     @staticmethod
     def tap_name():
@@ -40,7 +44,9 @@ class BaseTapTest(unittest.TestCase):
         return_value = {
             'start_date': '2017-07-01T00:00:00Z',
             'shop': 'stitchdatawearhouse',
-            'date_window_size': 30
+            'date_window_size': 30,
+            # BUG: https://jira.talendforge.org/browse/TDL-13180
+            # 'results_per_page': '50'
         }
 
         if original:
@@ -50,13 +56,20 @@ class BaseTapTest(unittest.TestCase):
         assert self.start_date > return_value["start_date"]
 
         return_value["start_date"] = self.start_date
+        return_value['shop'] = 'talenddatawearhouse'
         return return_value
 
     @staticmethod
-    def get_credentials():
+    def get_credentials(original_credentials: bool = True):
         """Authentication information for the test account"""
+
+        if original_credentials:
+            return {
+                'api_key': os.getenv('TAP_SHOPIFY_API_KEY_STITCHDATAWEARHOUSE')
+            }
+
         return {
-            'api_key': os.getenv('TAP_SHOPIFY_API_KEY')
+            'api_key': os.getenv('TAP_SHOPIFY_API_KEY_TALENDDATAWEARHOUSE')
         }
 
     def expected_metadata(self):
@@ -66,13 +79,18 @@ class BaseTapTest(unittest.TestCase):
                 self.REPLICATION_KEYS: {"updated_at"},
                 self.PRIMARY_KEYS: {"id"},
                 self.REPLICATION_METHOD: self.INCREMENTAL,
-                self.API_LIMIT: 250}
+                self.API_LIMIT: self.DEFAULT_RESULTS_PER_PAGE}
 
         meta = default.copy()
         meta.update({self.FOREIGN_KEYS: {"owner_id", "owner_resource"}})
 
         return {
-            "abandoned_checkouts": default,
+            "abandoned_checkouts": {
+                self.REPLICATION_KEYS: {"updated_at"},
+                self.PRIMARY_KEYS: {"id"},
+                self.REPLICATION_METHOD: self.INCREMENTAL,
+                # BUG: https://jira.talendforge.org/browse/TDL-13180
+                self.API_LIMIT: 50},
             "collects": default,
             "custom_collections": default,
             "smart_collections": default,
@@ -84,15 +102,35 @@ class BaseTapTest(unittest.TestCase):
                 self.REPLICATION_KEYS: {"created_at"},
                 self.PRIMARY_KEYS: {"id"},
                 self.REPLICATION_METHOD: self.INCREMENTAL,
-                self.API_LIMIT: 250},
+                self.API_LIMIT: self.DEFAULT_RESULTS_PER_PAGE},
             "products": default,
+            "inventory_items": {self.REPLICATION_KEYS: {"updated_at"},
+                self.PRIMARY_KEYS: {"id"},
+                self.REPLICATION_METHOD: self.INCREMENTAL,
+                self.API_LIMIT: 250},
             "metafields": meta,
             "transactions": {
                 self.REPLICATION_KEYS: {"created_at"},
                 self.PRIMARY_KEYS: {"id"},
                 self.FOREIGN_KEYS: {"order_id"},
                 self.REPLICATION_METHOD: self.INCREMENTAL,
-                self.API_LIMIT: 250}
+                self.API_LIMIT: self.DEFAULT_RESULTS_PER_PAGE},
+            "locations": {
+                self.REPLICATION_KEYS: {"updated_at"},
+                self.PRIMARY_KEYS: {"id"},
+                self.REPLICATION_METHOD: self.INCREMENTAL,
+                self.API_LIMIT: 0},
+            "inventory_levels": {
+                self.REPLICATION_KEYS: {"updated_at"},
+                self.PRIMARY_KEYS: {"location_id", "inventory_item_id"},
+                self.REPLICATION_METHOD: self.INCREMENTAL,
+                self.API_LIMIT: self.DEFAULT_RESULTS_PER_PAGE},
+            "events": {
+                self.REPLICATION_KEYS: {"created_at"},
+                self.PRIMARY_KEYS: {"id"},
+                self.REPLICATION_METHOD: self.INCREMENTAL,
+                self.API_LIMIT: 50
+            }
         }
 
     def expected_streams(self):
@@ -142,7 +180,10 @@ class BaseTapTest(unittest.TestCase):
 
     def setUp(self):
         """Verify that you have set the prerequisites to run the tap (creds, etc.)"""
-        missing_envs = [x for x in [os.getenv('TAP_SHOPIFY_API_KEY')] if x is None]
+        missing_envs = [x
+                        for x in [os.getenv('TAP_SHOPIFY_API_KEY_STITCHDATAWEARHOUSE'),
+                                  os.getenv('TAP_SHOPIFY_API_KEY_TALENDDATAWEARHOUSE')]
+                        if x is None]
         if missing_envs:
             raise Exception("set environment variables")
 
@@ -150,10 +191,10 @@ class BaseTapTest(unittest.TestCase):
     #   Helper Methods      #
     #########################
 
-    def create_connection(self, original_properties: bool = True):
+    def create_connection(self, original_properties: bool = True, original_credentials: bool = True):
         """Create a new connection with the test name"""
         # Create the connection
-        conn_id = connections.ensure_connection(self, original_properties)
+        conn_id = connections.ensure_connection(self, original_properties, original_credentials)
 
         # Run a check job using orchestrator (discovery)
         check_job_name = runner.run_check_mode(self, conn_id)
@@ -262,3 +303,56 @@ class BaseTapTest(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.start_date = self.get_properties().get("start_date")
+        self.store_1_streams = {'custom_collections', 'orders', 'products', 'customers', 'locations', 'inventory_levels', 'inventory_items', 'events'}
+        self.store_2_streams = {'abandoned_checkouts', 'collects', 'metafields', 'transactions', 'order_refunds', 'products', 'locations', 'inventory_levels', 'inventory_items', 'events'}
+
+    #modified this method to accommodate replication key in the current_state
+    def calculated_states_by_stream(self, current_state):
+        timedelta_by_stream = {stream: [0,5,0]  # {stream_name: [days, hours, minutes], ...}
+                               for stream in self.expected_streams()}
+
+        stream_to_calculated_state = {stream: "" for stream in current_state['bookmarks'].keys()}
+
+        for stream, state in current_state['bookmarks'].items():
+            # modified state value to accommodate the replication key
+            stream = stream
+            state_key, state_value = list(state.items())[0]
+
+            state_as_datetime = dateutil.parser.parse(list(state.values())[0])
+
+            days, hours, minutes = timedelta_by_stream[stream]
+            calculated_state_as_datetime = state_as_datetime - timedelta(days=days, hours=hours, minutes=minutes)
+
+            state_format = "%Y-%m-%dT00:00:00Z"
+            calculated_state_formatted = dt.strftime(calculated_state_as_datetime, state_format)
+            state = {state_key: calculated_state_formatted}
+            stream_to_calculated_state[stream] = state
+
+        return stream_to_calculated_state
+
+    def convert_state_to_utc(self, date_str):
+        """
+        Convert a saved bookmark value of the form '2020-08-25T13:17:36-07:00' to
+        a string formatted utc datetime,
+        in order to compare aginast json formatted datetime values
+        """
+        date_object = dateutil.parser.parse(date_str)
+        date_object_utc = date_object.astimezone(tz=pytz.UTC)
+        return dt.strftime(date_object_utc, "%Y-%m-%dT%H:%M:%SZ")
+
+    def timedelta_formatted(self, dtime, days=0):
+        try:
+            date_stripped = dt.strptime(dtime, self.START_DATE_FORMAT)
+            return_date = date_stripped + timedelta(days=days)
+
+            return dt.strftime(return_date, self.START_DATE_FORMAT)
+
+        except ValueError:
+            try:
+                date_stripped = dt.strptime(dtime, self.BOOKMARK_COMPARISON_FORMAT)
+                return_date = date_stripped + timedelta(days=days)
+
+                return dt.strftime(return_date, self.BOOKMARK_COMPARISON_FORMAT)
+
+            except ValueError:
+                return Exception("Datetime object is not of the format: {}".format(self.START_DATE_FORMAT))